reorganize screen (according to figma slices) 2
