+++ lib/screens/home_and_inner_pages/home_shell/blue_background.dart
