+++ lib/screens/kyc_verification_pages/kyc_verification_step1.dart
@@ -1,5 +1,5 @@
 import 'package:eto_pay/providers/kyc_form_provider.dart';
-import 'package:eto_pay/screens/kyc_verification_step2.dart';
+import 'package:eto_pay/screens/kyc_verification_pages/kyc_verification_step2.dart';
 import 'package:eto_pay/widgets/continue_button.dart';
 import 'package:eto_pay/widgets/custom_wide_input_field.dart';
 import 'package:eto_pay/widgets/progress_bar.dart';
