+++ lib/screens/kyc_verification_pages/kyc_verification_step2.dart
@@ -1,13 +1,13 @@
-import 'package:eto_pay/screens/kyc_verification_step1.dart';
-import 'package:eto_pay/screens/kyc_verification_step3.dart';
+import 'package:eto_pay/screens/kyc_verification_pages/kyc_verification_step1.dart';
+import 'package:eto_pay/screens/kyc_verification_pages/kyc_verification_step3.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:eto_pay/core/eu_countries.dart';
 import 'package:eto_pay/providers/kyc_form_provider.dart';
-import 'package:eto_pay/screens/kyc_verification_step2_driving_license.dart';
-import 'package:eto_pay/screens/kyc_verification_step2_id_card.dart';
-import 'package:eto_pay/screens/kyc_verification_step2_passport.dart';
+import 'package:eto_pay/screens/kyc_verification_pages/kyc_verification_step2_driving_license.dart';
+import 'package:eto_pay/screens/kyc_verification_pages/kyc_verification_step2_id_card.dart';
+import 'package:eto_pay/screens/kyc_verification_pages/kyc_verification_step2_passport.dart';
 import 'package:eto_pay/widgets/continue_button.dart';
 import 'package:eto_pay/widgets/country_dropdown.dart';
 import 'package:eto_pay/widgets/progress_bar.dart';
