+++ lib/screens/kyc_verification_pages/kyc_verification_step2_driving_license.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:eto_pay/core/eu_countries.dart';
 import 'package:eto_pay/providers/kyc_form_provider.dart';
-import 'package:eto_pay/screens/kyc_verification_step2.dart';
+import 'package:eto_pay/screens/kyc_verification_pages/kyc_verification_step2.dart';
 import 'package:eto_pay/widgets/continue_button.dart';
 import 'package:eto_pay/widgets/country_dropdown.dart';
 import 'package:eto_pay/widgets/image_upload_card.dart';
@@ -12,8 +12,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-class KycVerificationStep2PassportScreen extends ConsumerWidget {
-  const KycVerificationStep2PassportScreen({super.key});
+class KycVerificationStep2DrivingLicenseScreen extends ConsumerWidget {
+  const KycVerificationStep2DrivingLicenseScreen({super.key});
 
   void _updateIssuer(WidgetRef ref, String countryName) {
     final notifier = ref.read(kycFormProvider.notifier);
@@ -27,11 +27,7 @@ class KycVerificationStep2PassportScreen extends ConsumerWidget {
     required File? phoneFile,
   }) {
     final notifier = ref.read(kycFormProvider.notifier);
-    notifier.updatePassportImage(
-      isFront: isFront,
-      webBytes: webBytes,
-      phoneFile: phoneFile,
-    );
+    notifier.updateDrivingLicenseImage(isFront: isFront, webBytes: webBytes, phoneFile: phoneFile);
   }
 
   @override
@@ -103,19 +99,19 @@ class KycVerificationStep2PassportScreen extends ConsumerWidget {
                     ),
                     const SizedBox(height: 24),
                     const Text(
-                      "Passport",
+                      "Driving License",
                       style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
                     ),
                     const SizedBox(height: 8),
                     const Text(
-                      "Please upload clear images of both sides of your passport",
+                      "Please upload clear images of both sides of your driving license",
                       style: TextStyle(color: Color(0xFF747474), fontSize: 16),
                     ),
                     const SizedBox(height: 16),
                     ImageUploadCard(
                       label: "Front side",
-                      initialWebImage: form.passport?.front?.web,
-                      initialLocalImage: form.passport?.front?.phone,
+                      initialWebImage: form.drivingLicense?.front?.web,
+                      initialLocalImage: form.drivingLicense?.front?.phone,
                       onImageSelected: (bytes, file) => _updateDocumentImage(
                         ref: ref,
                         isFront: true,
@@ -126,8 +122,8 @@ class KycVerificationStep2PassportScreen extends ConsumerWidget {
                     const SizedBox(height: 24),
                     ImageUploadCard(
                       label: "Back side",
-                      initialWebImage: form.passport?.back?.web,
-                      initialLocalImage: form.passport?.back?.phone,
+                      initialWebImage: form.drivingLicense?.back?.web,
+                      initialLocalImage: form.drivingLicense?.back?.phone,
                       onImageSelected: (bytes, file) => _updateDocumentImage(
                         ref: ref,
                         isFront: false,
@@ -141,9 +137,9 @@ class KycVerificationStep2PassportScreen extends ConsumerWidget {
               ),
             ),
             ContinueButtonWidget(
-              isEnabled: form.isStep2PassportValid,
+              isEnabled: form.isStep2DrivingLicenseValid,
               text: 'Submit',
-              onPressed: form.isStep2PassportValid
+              onPressed: form.isStep2DrivingLicenseValid
                   ? () {
                       Navigator.of(context).push(
                         MaterialPageRoute(builder: (_) => const KycVerificationStep2Screen()),
