+++ lib/screens/kyc_verification_pages/kyc_verification_step2_passport.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:eto_pay/core/eu_countries.dart';
 import 'package:eto_pay/providers/kyc_form_provider.dart';
-import 'package:eto_pay/screens/kyc_verification_step2.dart';
+import 'package:eto_pay/screens/kyc_verification_pages/kyc_verification_step2.dart';
 import 'package:eto_pay/widgets/continue_button.dart';
 import 'package:eto_pay/widgets/country_dropdown.dart';
 import 'package:eto_pay/widgets/image_upload_card.dart';
@@ -12,8 +12,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-class KycVerificationStep2IdCardScreen extends ConsumerWidget {
-  const KycVerificationStep2IdCardScreen({super.key});
+class KycVerificationStep2PassportScreen extends ConsumerWidget {
+  const KycVerificationStep2PassportScreen({super.key});
 
   void _updateIssuer(WidgetRef ref, String countryName) {
     final notifier = ref.read(kycFormProvider.notifier);
@@ -27,8 +27,11 @@ class KycVerificationStep2IdCardScreen extends ConsumerWidget {
     required File? phoneFile,
   }) {
     final notifier = ref.read(kycFormProvider.notifier);
-    notifier.updateIdCardImage(
-        isFront: isFront, webBytes: webBytes, phoneFile: phoneFile);
+    notifier.updatePassportImage(
+      isFront: isFront,
+      webBytes: webBytes,
+      phoneFile: phoneFile,
+    );
   }
 
   @override
@@ -47,8 +50,7 @@ class KycVerificationStep2IdCardScreen extends ConsumerWidget {
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
                     TextButton(
-                      style:
-                          TextButton.styleFrom(foregroundColor: Colors.black),
+                      style: TextButton.styleFrom(foregroundColor: Colors.black),
                       onPressed: () => Navigator.of(context).pop(),
                       child: const Row(
                         mainAxisSize: MainAxisSize.min,
@@ -64,19 +66,16 @@ class KycVerificationStep2IdCardScreen extends ConsumerWidget {
                     const SizedBox(height: 20),
                     const Text(
                       "Step 2/4",
-                      style:
-                          TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
+                      style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
                     ),
                     const SizedBox(height: 20),
                     Row(
                       children: [
-                        SvgPicture.asset('assets/icons/icon_face_id.svg',
-                            width: 20, height: 20),
+                        SvgPicture.asset('assets/icons/icon_face_id.svg', width: 20, height: 20),
                         const SizedBox(width: 8),
                         const Text(
                           "2. ID Verification",
-                          style: TextStyle(
-                              fontWeight: FontWeight.bold, fontSize: 18),
+                          style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18),
                         ),
                       ],
                     ),
@@ -88,15 +87,13 @@ class KycVerificationStep2IdCardScreen extends ConsumerWidget {
                     const SizedBox(height: 42),
                     const Text(
                       "Select document issuer country",
-                      style:
-                          TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
+                      style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
                     ),
                     const SizedBox(height: 8),
                     CountryDropdown(
                       countries: euCountries,
                       selectedCountry: form.idVerificationDocumentIssuer != null
-                          ? findCountryByName(
-                              form.idVerificationDocumentIssuer!)
+                          ? findCountryByName(form.idVerificationDocumentIssuer!)
                           : null,
                       onChanged: (value) {
                         if (value != null) {
@@ -106,20 +103,19 @@ class KycVerificationStep2IdCardScreen extends ConsumerWidget {
                     ),
                     const SizedBox(height: 24),
                     const Text(
-                      "ID Card",
-                      style:
-                          TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
+                      "Passport",
+                      style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
                     ),
                     const SizedBox(height: 8),
                     const Text(
-                      "Please upload clear images of both sides of your ID card.",
+                      "Please upload clear images of both sides of your passport",
                       style: TextStyle(color: Color(0xFF747474), fontSize: 16),
                     ),
                     const SizedBox(height: 16),
                     ImageUploadCard(
                       label: "Front side",
-                      initialWebImage: form.idCard?.front?.web,
-                      initialLocalImage: form.idCard?.front?.phone,
+                      initialWebImage: form.passport?.front?.web,
+                      initialLocalImage: form.passport?.front?.phone,
                       onImageSelected: (bytes, file) => _updateDocumentImage(
                         ref: ref,
                         isFront: true,
@@ -130,8 +126,8 @@ class KycVerificationStep2IdCardScreen extends ConsumerWidget {
                     const SizedBox(height: 24),
                     ImageUploadCard(
                       label: "Back side",
-                      initialWebImage: form.idCard?.back?.web,
-                      initialLocalImage: form.idCard?.back?.phone,
+                      initialWebImage: form.passport?.back?.web,
+                      initialLocalImage: form.passport?.back?.phone,
                       onImageSelected: (bytes, file) => _updateDocumentImage(
                         ref: ref,
                         isFront: false,
@@ -145,13 +141,12 @@ class KycVerificationStep2IdCardScreen extends ConsumerWidget {
               ),
             ),
             ContinueButtonWidget(
-              isEnabled: form.isStep2IdCardValid,
+              isEnabled: form.isStep2PassportValid,
               text: 'Submit',
-              onPressed: form.isStep2IdCardValid
+              onPressed: form.isStep2PassportValid
                   ? () {
                       Navigator.of(context).push(
-                        MaterialPageRoute(
-                            builder: (_) => const KycVerificationStep2Screen()),
+                        MaterialPageRoute(builder: (_) => const KycVerificationStep2Screen()),
                       );
                     }
                   : () {},
