+++ lib/screens/kyc_verification_pages/kyc_verification_step3.dart
@@ -1,6 +1,6 @@
 import 'package:eto_pay/providers/kyc_form_provider.dart';
-import 'package:eto_pay/screens/kyc_verification_step3_upload_photo.dart';
-import 'package:eto_pay/screens/kyc_verification_step4.dart';
+import 'package:eto_pay/screens/kyc_verification_pages/kyc_verification_step3_upload_photo.dart';
+import 'package:eto_pay/screens/kyc_verification_pages/kyc_verification_step4.dart';
 import 'package:eto_pay/widgets/continue_button.dart';
 import 'package:eto_pay/widgets/progress_bar.dart';
 import 'package:eto_pay/widgets/wide_button_with_icon_and_arrow.dart';
