+++ lib/screens/profile_creation_pages/upload_recovery_file_screen.dart
@@ -1,5 +1,5 @@
 import 'dart:typed_data';
-import 'package:eto_pay/screens/upload_recovery_file_import_file_screen.dart';
+import 'package:eto_pay/screens/profile_creation_pages/upload_recovery_file_import_file_screen.dart';
 import 'package:eto_pay/widgets/dashed_container.dart';
 import 'package:flutter/material.dart';
 import 'package:eto_pay/core/colors.dart';
