+++ lib/views/recovery_phrase_view.dart
@@ -3,7 +3,7 @@ import 'package:eto_pay/view_models/recovery_phrase_view_model.dart';
 import 'package:eto_pay/widgets/conditional_button.single.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:eto_pay/screens/recovery_phrase_verification_screen.dart';
+import 'package:eto_pay/screens/profile_creation_pages/recovery_phrase_verification_screen.dart';
 import 'package:eto_pay/core/colors.dart';
 
 class RecoveryPhraseView extends StatelessWidget {
