+++ test/create_profile_screen_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
-import 'package:eto_pay/screens/create_profile_screen.dart';
+import 'package:eto_pay/screens/profile_creation_pages/create_profile_screen.dart';
 
 void main() {
   testWidgets('CreateProfileScreen renders and shows software profile card',
