+++ test/pin_setup_screen_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
-import 'package:eto_pay/screens/pin_setup_screen.dart';
+import 'package:eto_pay/screens/profile_creation_pages/pin_setup_screen.dart';
 
 void main() {
   testWidgets('PinSetupScreen renders and shows pin fields',
