+++ test/recovery_phrase_verification_screen_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
-import 'package:eto_pay/screens/recovery_phrase_verification_screen.dart';
+import 'package:eto_pay/screens/profile_creation_pages/recovery_phrase_verification_screen.dart';
 
 void main() {
   testWidgets('RecoveryPhraseVerificationScreen renders and shows Next button',
